Start Button Image
Sorry I forgot to include this sprite the other day when I updated the code. Over time, for this port, this sprite will almost certainly change. I will be soon implementing the scaling code for Hero, and I'll also design the start menu. This sprite was just taken from the Scratch project.